More test coverage of MQTT etc
